requirements.txt
# Habit Tracker Application Dependencies
# Install with: pip install -r requirements.txt

Flask==2.3.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-CORS==4.0.0
Authlib==1.3.0
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0
# PostgreSQL adapter - may require Visual C++ Build Tools on Windows
psycopg2-binary==2.9.7
python-decouple==3.8

# Testing dependencies
pytest==7.4.3
hypothesis==6.88.1
psutil==5.9.6